Moved text_button class into a new file "text_button.py"
